# docs/api/integrated-api-openapi.json
{
    "components": {
        "schemas": {
            "ErrorResponse": {
                "description": "\uacf5\ud1b5 \uc5d0\ub7ec \uc751\ub2f5",
                "properties": {
                    "code": {
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc",
                        "example": "BE40101",
                        "type": "string"
                    },
                    "message": {
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0",
                        "example": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc740 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                        "type": "string"
                    },
                    "status": {
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 401,
                        "format": "int32",
                        "type": "integer"
                    },
                    "timestamp": {
                        "description": "\uc5d0\ub7ec \ubc1c\uc0dd \uc2dc\uac01",
                        "format": "date-time",
                        "type": "string"
                    }
                }
            },
...
            "com.backend.commondataaccess.dto.OffsetPageResultCom.backend.integratedapi.collectsource.controller.dto.CollectSourceReadDto.Response": {
                "description": "\uc624\ud504\uc14b \uae30\ubc18 \ud398\uc774\uc9d5 \uc751\ub2f5",
                "properties": {
                    "items": {
                        "description": "\ud604\uc7ac \ud398\uc774\uc9c0 \ud56d\ubaa9 \ubaa9\ub85d",
                        "items": {
                            "$ref": "#/components/schemas/com.backend.integratedapi.collectsource.controller.dto.CollectSourceReadDto.Response"
                        },
                        "type": "array"
                    },
                    "page": {
                        "description": "\ud604\uc7ac \ud398\uc774\uc9c0 \ubc88\ud638(0\ubd80\ud130 \uc2dc\uc791)",
                        "example": 0,
                        "format": "int32",
                        "type": "integer"
                    },
                    "size": {
                        "description": "\ud398\uc774\uc9c0 \ud06c\uae30",
                        "example": 20,
                        "format": "int32",
                        "type": "integer"
                    },
                    "total_count": {
                        "description": "\uc804\uccb4 \uac74\uc218",
                        "example": 95,
                        "format": "int64",
                        "type": "integer"
                    }
                }
            },
...
            "com.backend.integratedapi.collectingjob.controller.dto.CollectingJobReadDto.Response": {
                "description": "\uc218\uc9d1 Job \uc870\ud68c \uacb0\uacfc",
                "properties": {
                    "attempt_count": {
                        "description": "\uc2dc\ub3c4 \ud69f\uc218",
                        "format": "int32",
                        "type": "integer"
                    },
                    "collected_count": {
                        "description": "\uc2e4\uc81c \uc218\uc9d1\ub41c \uac74\uc218",
                        "format": "int32",
                        "type": "integer"
                    },
                    "collecting_status": {
                        "description": "\uc218\uc9d1 \uc9c4\ud589 \uc0c1\ud0dc",
                        "enum": [
                            "discovered",
                            "fetched",
                            "fetch_failed",
                            "parsed",
                            "parse_failed"
                        ],
                        "type": "string"
                    },
                    "ended_at": {
                        "description": "\uc885\ub8cc \uc2dc\uac01",
                        "format": "date-time",
                        "type": "string"
                    },
                    "error_message": {
                        "description": "\uc2e4\ud328 \uc2dc \uc5d0\ub7ec \uba54\uc2dc\uc9c0",
                        "type": "string"
                    },
                    "force_recollect": {
                        "description": "\uac15\uc81c \uc7ac\uc218\uc9d1 \uc5ec\ubd80",
                        "type": "boolean"
                    },
                    "from_page": {
                        "description": "\uc218\uc9d1 \uc2dc\uc791 \ud398\uc774\uc9c0",
                        "format": "int32",
                        "type": "integer"
                    },
                    "job_id": {
                        "description": "Job ID",
                        "format": "uuid",
                        "type": "string"
                    },
                    "job_status": {
                        "description": "Job \uc0c1\ud0dc",
                        "enum": [
                            "pending",
                            "running",
                            "success",
                            "failed",
                            "cancelled"
                        ],
                        "type": "string"
                    },
                    "started_at": {
                        "description": "\uc2dc\uc791 \uc2dc\uac01",
                        "format": "date-time",
                        "type": "string"
                    },
                    "to_page": {
                        "description": "\uc218\uc9d1 \uc885\ub8cc \ud398\uc774\uc9c0",
                        "format": "int32",
                        "type": "integer"
                    },
                    "total_count": {
                        "description": "\uc218\uc9d1 \ub300\uc0c1 \uc804\uccb4 \uac74\uc218",
                        "format": "int32",
                        "type": "integer"
                    },
                    "triggered_by": {
                        "description": "\uc2e4\ud589 \uc694\uccad\uc790 \uc0ac\uc6a9\uc790 ID",
                        "format": "uuid",
                        "type": "string"
                    }
                }
            },
...
            "com.backend.integratedapi.collectsource.controller.dto.CollectSourceCreateDto.Response": {
                "description": "\uc218\uc9d1 \uc18c\uc2a4 \uc0dd\uc131 \uacb0\uacfc",
                "properties": {
                    "created_at": {
                        "description": "\uc0dd\uc131 \uc2dc\uac01",
                        "format": "date-time",
                        "type": "string"
                    },
                    "source_id": {
                        "description": "\uc0dd\uc131\ub41c \uc218\uc9d1 \uc18c\uc2a4 ID",
                        "format": "uuid",
                        "type": "string"
                    }
                }
            },
...
            "com.backend.integratedapi.collectsourcepost.controller.dto.CollectSourcePostReadDto.Response": {
                "description": "\uc218\uc9d1\ub41c \uac8c\uc2dc\uae00 \uc870\ud68c \uacb0\uacfc",
                "properties": {
                    "collect_source_id": {
                        "description": "\uc218\uc9d1 \uc18c\uc2a4 ID",
                        "format": "uuid",
                        "type": "string"
                    },
                    "indexing_error_count": {
                        "description": "\uc0c9\uc778 \uc2e4\ud328 \ub204\uc801 \ud69f\uc218",
                        "format": "int32",
                        "type": "integer"
                    },
                    "indexing_status": {
                        "description": "\uc0c9\uc778 \uc0c1\ud0dc",
                        "enum": [
                            "pending",
                            "indexing",
                            "indexed",
                            "failed",
                            "skipped"
                        ],
                        "type": "string"
                    },
                    "last_collected_at": {
                        "description": "\ub9c8\uc9c0\ub9c9 \uc218\uc9d1 \uc2dc\uac01",
                        "format": "date-time",
                        "type": "string"
                    },
                    "last_collecting_job_id": {
                        "description": "\ub9c8\uc9c0\ub9c9 \uc218\uc9d1 Job ID",
                        "format": "uuid",
                        "type": "string"
                    },
                    "last_indexed_at": {
                        "description": "\ub9c8\uc9c0\ub9c9 \uc0c9\uc778 \uc2dc\uac01",
                        "format": "date-time",
                        "type": "string"
                    },
                    "posting_id": {
                        "description": "\uac8c\uc2dc\uae00 ID",
                        "format": "uuid",
                        "type": "string"
                    },
                    "published_at": {
                        "description": "\uc6d0\ubb38 \ubc1c\ud589\uc77c",
                        "format": "date",
                        "type": "string"
                    },
                    "summary": {
                        "description": "\ubcf8\ubb38 \uc694\uc57d",
                        "type": "string"
                    },
                    "thumbnail_url": {
                        "description": "\uc378\ub124\uc77c \uc774\ubbf8\uc9c0 URL",
                        "type": "string"
                    },
                    "title": {
                        "description": "\uac8c\uc2dc\uae00 \uc81c\ubaa9",
                        "type": "string"
                    },
                    "url": {
                        "description": "\uc6d0\ubb38 URL",
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "bearerFormat": "JWT",
                "description": "user-service \ub85c\uadf8\uc778 \ub610\ub294 \uad6c\uae00 OAuth \ucf5c\ubc31\uc5d0\uc11c \ubc1c\uae09\ubc1b\uc740 access token",
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "integrated-service REST API",
        "title": "Integrated-Service API",
        "version": "v1"
    },
    "openapi": "3.1.0",
    "paths": {
...
        "/collect/v1/jobs/{id}": {
            "get": {
                "description": "\uc6b4\uc601\uc790 \uc804\uc6a9 API",
                "operationId": "getSource_2",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/com.backend.integratedapi.collectingjob.controller.dto.CollectingJobReadDto.Response"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "FE40001",
                                    "message": "\uc785\ub825 \ub370\uc774\ud130\uc5d0 \ubb38\uc81c\uac00 \uc788\uc2b5\ub2c8\ub2e4.",
                                    "status": 400,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\uc694\uccad \uac12\uc774 \uc62c\ubc14\ub974\uc9c0 \uc54a\uc74c"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "BE40101",
                                    "message": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc740 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                                    "status": 401,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \uc720\ud6a8\ud558\uc9c0 \uc54a\uc74c"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "BE40301",
                                    "message": "\uad8c\ud55c\uc774 \uc5c6\ub294 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                                    "status": 403,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\uad8c\ud55c\uc774 \uc5c6\ub294 \uc0ac\uc6a9\uc790"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "FE50001",
                                    "message": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.",
                                    "status": 500,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\ucc98\ub9ac\ub418\uc9c0 \uc54a\uc740 \uc11c\ubc84 \uc624\ub958"
                    }
                },
                "summary": "\uc218\uc9d1 \uc791\uc5c5 \uc0c1\ud0dc \uc870\ud68c",
                "tags": [
                    "03. \uc218\uc9d1 \uc791\uc5c5 \uc2e4\ud589 API"
                ]
            }
        },
        "/collect/v1/postings/{id}": {
            "get": {
                "description": "\uc6b4\uc601\uc790 \uc804\uc6a9 API",
                "operationId": "getSource_1",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/com.backend.integratedapi.collectsourcepost.controller.dto.CollectSourcePostReadDto.Response"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "FE40001",
                                    "message": "\uc785\ub825 \ub370\uc774\ud130\uc5d0 \ubb38\uc81c\uac00 \uc788\uc2b5\ub2c8\ub2e4.",
                                    "status": 400,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\uc694\uccad \uac12\uc774 \uc62c\ubc14\ub974\uc9c0 \uc54a\uc74c"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "BE40101",
                                    "message": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc740 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                                    "status": 401,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \uc720\ud6a8\ud558\uc9c0 \uc54a\uc74c"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "BE40301",
                                    "message": "\uad8c\ud55c\uc774 \uc5c6\ub294 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                                    "status": 403,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\uad8c\ud55c\uc774 \uc5c6\ub294 \uc0ac\uc6a9\uc790"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "FE50001",
                                    "message": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.",
                                    "status": 500,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\ucc98\ub9ac\ub418\uc9c0 \uc54a\uc740 \uc11c\ubc84 \uc624\ub958"
                    }
                },
                "summary": "\uc218\uc9d1 \uacb0\uacfc(\uc6d0\ubb38/\uba54\ud0c0) \uc870\ud68c",
                "tags": [
                    "04. \uc218\uc9d1 \uacb0\uacfc \uc870\ud68c API"
                ]
            }
        },
...
        "/collect/v1/sources": {
            "get": {
                "description": "\uc6b4\uc601\uc790 \uc804\uc6a9 API",
                "operationId": "getSources",
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "required": false,
                        "schema": {
                            "default": 0,
                            "format": "int32",
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "size",
                        "required": false,
                        "schema": {
                            "default": 10,
                            "format": "int32",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/com.backend.commondataaccess.dto.OffsetPageResultCom.backend.integratedapi.collectsource.controller.dto.CollectSourceReadDto.Response"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "FE40001",
                                    "message": "\uc785\ub825 \ub370\uc774\ud130\uc5d0 \ubb38\uc81c\uac00 \uc788\uc2b5\ub2c8\ub2e4.",
                                    "status": 400,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\uc694\uccad \uac12\uc774 \uc62c\ubc14\ub974\uc9c0 \uc54a\uc74c"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "BE40101",
                                    "message": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc740 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                                    "status": 401,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \uc720\ud6a8\ud558\uc9c0 \uc54a\uc74c"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "BE40301",
                                    "message": "\uad8c\ud55c\uc774 \uc5c6\ub294 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                                    "status": 403,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\uad8c\ud55c\uc774 \uc5c6\ub294 \uc0ac\uc6a9\uc790"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "FE50001",
                                    "message": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.",
                                    "status": 500,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\ucc98\ub9ac\ub418\uc9c0 \uc54a\uc740 \uc11c\ubc84 \uc624\ub958"
                    }
                },
                "summary": "\uc218\uc9d1 \uc18c\uc2a4 \ubaa9\ub85d \uc870\ud68c",
                "tags": [
                    "02. \uc218\uc9d1 \uc18c\uc2a4 \uad00\ub9ac API"
                ]
            },
            "post": {
                "description": "\uc6b4\uc601\uc790 \uc804\uc6a9 API",
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/com.backend.integratedapi.collectsource.controller.dto.CollectSourceCreateDto.Request"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/com.backend.integratedapi.collectsource.controller.dto.CollectSourceCreateDto.Response"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "FE40001",
                                    "message": "\uc785\ub825 \ub370\uc774\ud130\uc5d0 \ubb38\uc81c\uac00 \uc788\uc2b5\ub2c8\ub2e4.",
                                    "status": 400,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\uc694\uccad \uac12\uc774 \uc62c\ubc14\ub974\uc9c0 \uc54a\uc74c"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "BE40101",
                                    "message": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc740 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                                    "status": 401,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \uc720\ud6a8\ud558\uc9c0 \uc54a\uc74c"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "BE40301",
                                    "message": "\uad8c\ud55c\uc774 \uc5c6\ub294 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                                    "status": 403,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\uad8c\ud55c\uc774 \uc5c6\ub294 \uc0ac\uc6a9\uc790"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "FE50001",
                                    "message": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.",
                                    "status": 500,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\ucc98\ub9ac\ub418\uc9c0 \uc54a\uc740 \uc11c\ubc84 \uc624\ub958"
                    }
                },
                "summary": "\uc218\uc9d1 \uc18c\uc2a4 \ub4f1\ub85d",
                "tags": [
                    "02. \uc218\uc9d1 \uc18c\uc2a4 \uad00\ub9ac API"
                ]
            }
        },
...
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "servers": [
        {
            "description": "Generated server url",
            "url": "http://localhost:8081"
        }
    ],
...
}